combine restart and hdf5 output into one code path
